apply fewshots; refactor
